in the main, switch back to Ido from Helm
Helm's find files dialogue often winds me up, so it's probably better to
get into the habit of opening up a bunch of dired buffers and making use
of all of dired's features.

More importantly there is the issue I tried to resolve by advising
`helm-completing-read-default-1`, but never fully got around because of
fundamentals of Helm's design.  It's too hard to make Helm make the
right choice of what will be selected when I hit enter, that is, the
choice between my input exactly and the closest match to my input.