`projectile-find-file-dwim' replaced with non-dwim
When I change to a project I'm presented with dired in the project's
root directory.  Point is generally on a subdirectory, so hitting C-c j
will then only let me jump to files in that subdirectory.  That's pretty
annoying.

Could advise projectile to move point when opening a dired buffer instead.