rework git-push-all
This avoids failure in the case where more than one remote has a branch
of the same name, but the local branch of that name can be pushed to
only one of them (and indeed, has its pushRemote set).

E.g. the master branch at origin and locally is the Debian packaging
branch, but the GitHub fork (used for submitting pull requests) has a
master branch that is upstream's master branch.  pushRemote for master
is set to origin, but since another branch has pushRemote set to fork,
the old script would try to push all matching branches to fork, which
fails as it can't push master there (not a fast-forward).