replace org-mode flagged snippet with a capture template
This is needed because notmuch thread ids are different on different
machines, so links created by the old snippet would work on only the
machine upon which they were created.

This new search does the trick and can be a capture template rather
than a yasnippet, which is much more convenient anyway.